refactor: Reorganize file structure
- Remake folders as Application, Presentation, Domain, Resources
- Extract use-case code from APIService.swift into MovieAPI.swift
- Rename Extension.swift to StyleExtensions.swift (Need to rework in future)
- Rename Constants.swift to Identifiers.swift (Need to remove after issue #42)
