+++ MovieApp/MovieApp/Domain/Entities/GenreModel.swift
