+++ MovieApp/MovieApp/Domain/Entities/MovieDetail.swift
