+++ MovieApp/MovieApp/Domain/Entities/MovieList.swift
