+++ MovieApp/MovieApp/Resources/Assets.xcassets/AppIcon.appiconset/40.png
