+++ MovieApp/MovieApp/Resources/Assets.xcassets/Color Set/LightBgColor.colorset/Contents.json
