+++ MovieApp/MovieApp/Resources/Assets.xcassets/Color Set/PlaceholderTextColor.colorset/Contents.json
