+++ MovieApp/MovieApp/Resources/Assets.xcassets/img_placeholder.imageset/Contents.json
